U.Codebase.Branch.Type

Documentation

data Branch m

A re-imagining of Unison.Codebase.Branch which is less eager in what it loads, which can often speed up load times and keep fewer things in memory.

Constructors

Branch

Fields

terms :: !(Map NameSegment (Map Referent (m MdValues)))

types :: !(Map NameSegment (Map Reference (m MdValues)))

patches :: !(Map NameSegment (PatchHash, m Patch))

children :: !(Map NameSegment (CausalBranch m))

type CausalBranch m = Causal m CausalHash BranchHash (Branch m) (Branch m)

data Patch

Constructors

Patch

Fields

termEdits :: !(Map Referent (Set TermEdit))

typeEdits :: !(Map Reference (Set TypeEdit))

type MetadataType = Reference

type MetadataValue = Reference

newtype MdValues

Constructors

MdValues

Fields

unMdValues :: Set MetadataValue

data NamespaceStats

Useful statistics about a namespace. All contained statistics should be static, i.e. they can be computed when a branch is first saved, and won't change unless the branch hash also changes.

Constructors

NamespaceStats

Fields

numContainedTerms :: !Int

numContainedTypes :: !Int

numContainedPatches :: !Int

hasDefinitions :: NamespaceStats -> Bool

Whether the provided stats indicate the presence of any definitions in the namespace.

childAt :: NameSegment -> Branch m -> Maybe (CausalBranch m)

hoist :: Functor n => (forall x. m x -> n x) -> Branch m -> Branch n

hoistCausalBranch :: Functor n => (forall x. m x -> n x) -> CausalBranch m -> CausalBranch n

empty :: Branch m