U.Codebase.Decl

Documentation

type ConstructorId = Word64

data DeclType

Constructors

Data

Effect

type Decl v = DeclR TypeRef v

type HashableDecl v = DeclR HashableTypeRef v

type TypeRef = Reference' Text (Maybe Hash)

type HashableTypeRef = Reference' Text Hash

type Type v = TypeR TypeRef v

type HashableType v = TypeR HashableTypeRef v

data Modifier

Constructors

Structural

Unique Text

data DeclR r v

Constructors

DataDeclaration

Fields

declType :: DeclType

modifier :: Modifier

bound :: [v]

constructorTypes :: [TypeR r v]

allVars :: Ord v => DeclR r v -> Set v

vmap :: Ord v' => (v -> v') -> DeclR r v -> DeclR r v'

rmap :: Ord v => (r -> r') -> DeclR r v -> DeclR r' v

Hashing stuff

dependencies :: (Ord r, Ord v) => DeclR r v -> Set r

data V v

Constructors

Bound v

Ctor Int

data F a

Constructors

Type (FD a)

LetRec [a] a

Constructors [a]

Modified DeclType Modifier a

unhashComponent

Arguments

:: forall v extra. Var v

=> Hash

-> (Id -> v)
A function to convert a reference to a variable. The actual var names aren't important.

-> Map Id (Decl v, extra)

-> Map Id (v, HashableDecl v, extra)
The component with all self-references replaced with variable references.

Given the pieces of a single decl component, replaces all Nothing self-referential hashes with a variable reference to the relevant piece of the component in the component map.