U.Codebase.Term

Documentation

type ConstructorId = Word64

type Term v = Term (F v) v ()

type HashableTerm v = Term (F' Text HashableTermRef TypeRef HashableTermLink TypeLink v) v ()

A version of Term but where TermRefs never have a Nothing Hash, but instead self references are filled with User Variable references to the relevant piece of the component in a component map.

type Type v = TypeR TypeRef v

type TermRef = Reference' Text (Maybe Hash)

type HashableTermRef = Reference' Text Hash

type TypeRef = Reference

type TermLink = Referent' (Reference' Text (Maybe Hash)) (Reference' Text Hash)

type HashableTermLink = Referent' (Reference' Text Hash) (Reference' Text Hash)

type TypeLink = Reference

type F vt = F' Text TermRef TypeRef TermLink TypeLink vt

Base functor for terms in the Unison codebase

data F' text termRef typeRef termLink typeLink vt a

Generalized version. We could generalize further to allow sharing within terms.

Constructors

Int Int64

Nat Word64

Float Double

Boolean Bool

Text text

Char Char

Ref termRef

Constructor typeRef ConstructorId

Request typeRef ConstructorId

Handle a a

App a a

Ann a (TypeR typeRef vt)

List (Seq a)

If a a a

And a a

Or a a

Lam a

LetRec [a] a

Let a a

Match a [MatchCase text typeRef a]

TermLink termLink

TypeLink typeLink

data MatchCase t r a

Constructors

MatchCase (Pattern t r) (Maybe a) a

data Pattern t r

Constructors

PUnbound

PVar

PBoolean !Bool

PInt !Int64

PNat !Word64

PFloat !Double

PText !t

PChar !Char

PConstructor !r !ConstructorId [Pattern t r]

PAs (Pattern t r)

PEffectPure (Pattern t r)

PEffectBind !r !ConstructorId [Pattern t r] (Pattern t r)

PSequenceLiteral [Pattern t r]

PSequenceOp (Pattern t r) !SeqOp (Pattern t r)

data SeqOp

Constructors

PCons

PSnoc

PConcat

extraMap :: forall text termRef typeRef termLink typeLink vt text' termRef' typeRef' termLink' typeLink' vt' v a. (Ord v, Ord vt') => (text -> text') -> (termRef -> termRef') -> (typeRef -> typeRef') -> (termLink -> termLink') -> (typeLink -> typeLink') -> (vt -> vt') -> Term (F' text termRef typeRef termLink typeLink vt) v a -> Term (F' text' termRef' typeRef' termLink' typeLink' vt') v a

extraMapM :: forall m text termRef typeRef termLink typeLink vt text' termRef' typeRef' termLink' typeLink' vt' v a. (Ord v, Ord vt', Monad m) => (text -> m text') -> (termRef -> m termRef') -> (typeRef -> m typeRef') -> (termLink -> m termLink') -> (typeLink -> m typeLink') -> (vt -> m vt') -> Term (F' text termRef typeRef termLink typeLink vt) v a -> m (Term (F' text' termRef' typeRef' termLink' typeLink' vt') v a)

rmapPattern :: (t -> t') -> (r -> r') -> Pattern t r -> Pattern t' r'

rmapPatternM :: Applicative m => (t -> m t') -> (r -> m r') -> Pattern t r -> m (Pattern t' r')

dependencies :: (Ord termRef, Ord typeRef, Ord termLink, Ord typeLink, Ord v) => Term (F' text termRef typeRef termLink typeLink vt) v a -> (Set termRef, Set typeRef, Set termLink, Set typeLink)

unhashComponent

Arguments

:: forall v extra. Var v

=> Hash
The hash of the component, this is used to fill in self-references.

-> (Id -> v)
A function to convert a reference to a variable. The actual var names aren't important.

-> Map Id (Term v, extra)

-> Map Id (v, HashableTerm v, extra)
The component with all self-references replaced with variable references.

Given the pieces of a single term component, replaces all Nothing self-referential hashes with a variable reference to the relevant piece of the component in the component map.